route配置，讓app.component只做navigate的工作
