+++ src/app/app.module.ts
@@ -2,8 +2,10 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { HttpModule } from '@angular/http';
+import { RouterModule } from '@angular/router';
 
 import { AppComponent } from './app.component';
+import { RawLessOneYearComponent } from './raw.less.one.year.component';
 
 
 @NgModule({
@@ -13,7 +15,13 @@ import { AppComponent } from './app.component';
   imports: [
     BrowserModule,
     FormsModule,
-    HttpModule
+    HttpModule,
+    RouterModule.forRoot([
+      {
+        path: 'raw-less-one-year',
+        component: RawLessOneYearComponent
+      }
+    ])
   ],
   providers: [],
   bootstrap: [AppComponent]
