+++ src/app/raw.less.one.year.component.css
@@ -1,0 +1,78 @@
+.title{
+    background-color: rgba(77,77,77,0.8);
+    color: #ccff99;
+    font-weight: 900;
+    max-width: 600px;
+    padding: 20px 40px;
+
+    border-radius: 10px;
+    box-shadow:  0px 0px 10px rgba(0,0,0,0.2), 0px 10px 20px -10px rgba(0,0,0,0.3);  
+}
+
+
+.page{
+    padding: 40px;
+    margin: 10px;
+    max-width: 600px;
+    background-color: #fff;
+    box-shadow: 0px 0px 10px rgba(0,0,0,0.3), 0px 25px 60px -10px rgba(0,0,0,0.3);
+    border-radius: 5px;    
+    color: #555;
+}
+
+
+.input-data{
+    font-size: 20px;
+    line-height: 40px;
+}
+
+
+input{
+    width: 80px;
+    border: solid 1px gray;
+    padding: 5px;
+    margin: 5px;
+    text-align: center;
+    font-size: 20px;
+}
+
+
+.memo p{
+    font-size: 16px;
+}
+
+
+.output-data{
+    font-size: 24px;
+}
+
+
+button{
+    border: solid 1px #1a3300;
+    border-radius: 2px;
+    box-shadow: 0px 0px 10px rgba(0,0,0,0.3);
+    background: #e6ffcc; /* For browsers that do not support gradients */
+
+    color: #1a3300;
+    background: rgba(#fff,0.9);
+    
+    width: 80px;
+    height: 40px;
+
+    margin: 10px;
+    padding: 6px;
+    cursor: pointer;
+
+    font-size: 20px;
+    font-weight: 900;
+    letter-spacing: 10px;
+    text-align: center;
+
+    transition-duration: 0.3s;
+}
+
+
+button:hover{
+    background:   #1a3300; /* For browsers that do not support gradients */
+    color: #e6ffcc;
+}
